双链表变成循环的了; 修复保存不了手机号的问题

+++ data_structure/chapter3/src/singlelinklist.cpp
@@ -6,11 +6,11 @@ using namespace std;
 
 struct Student
 {
-	int ID;
+	long ID;
 	string Name;
 	string Sex;
 	string Class;
-	int Phone;
+	long Phone;
 
 	bool operator<(const Student& other)
 	{
